Move cronjob package to pkg/
